pyproject.toml
[tool.poetry]
name = "clinical-speech"
version = "0.1.0"
description = "tba"
authors = ["Lasse <lasseh0310@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.12.1"
torchaudio = "^0.12.1"
numpy = "^1.23.3"
pytorch-lightning = "^1.7.7"
pandas = "^1.5.0"
flake8 = "^5.0.4"
black = "^22.10.0"
opensmile = "^2.4.1"
speechbrain = "^0.5.13"
torch-audiomentations = "^0.11.0"
scikit-learn = "^1.1.2"
wandb = "^0.13.4"
wasabi = "^0.10.1"
hydra-core = "^1.2.0"
transformers = "^4.23.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pylint]
load-plugins = "pylint.extensions.docparams,pylint.extensions.code_style,pylint.extensions.for_any_all,pylint.extensions.typing"
good-names = "df,p,f,d,e,n,k,i,v"
disable = "too-many-lines,line-too-long,missing-raises-doc,no-self-argument,unused-wildcard-import,wildcard-import,no-else-return,too-many-arguments,redefined-outer-name,c-extension-no-member,wrong-import-order"